Fixes and refactoring in the graph tests
