+++ src/test/java/ml/grafos/okapi/graphs/SemiClusteringTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ml.grafos.okapi.clustering;
+package ml.grafos.okapi.graphs;
 
 import static org.junit.Assert.*;
